requirements.txt
pymupdf4llm
pandas
google-genai
openai
python-dotenv
datasets 
pandas 
tqdm
asyncio
tiktoken
qdrant-client

numpy
langchain
langchain-text-splitters
pathlib2
groq
pymongo


streamlit
pydantic
fastapi
uvicorn
requests
aiohttp 
aiofiles